NGO/requirement.txt
asgiref==3.4.1
cffi==1.14.6
cryptography==3.4.7
Deprecated==1.2.12
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
jwcrypto==0.9.1
mysqlclient==2.0.3
pycparser==2.20
PyJWT==2.1.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
wrapt==1.12.1
